refactor : product code 위치와 동일하게 수정
