+++ src/test/java/cart/dao/member/JdbcMemberDaoTest.java
@@ -1,6 +1,5 @@
-package cart.dao;
+package cart.dao.member;
 
-import cart.dao.member.MemberDao;
 import cart.entity.MemberEntity;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
